Move routing to e1, bridging to proxy
Since we will be implementing a proxy in the future, we don't want
to mix its logic with routing decisions. It's easier to have the
routing on a separate node (here, one of the emulation nodes).

Bridging is always configured on the proxy and on the other (e2)
emulation node.