Add setup time parameter to network setup in tests
These mininet tests can be somewhat flaky. I think it has to do with
attempting to run mininet commands before the network is fully setup.
In general, these tests take a while to run (a few seconds each). It
would be nice if we could decrease the length of the feedback loop.