Fix flaky callback tests!
"Background processes" actually involved both a background process
for executing the requested command, and a background thread for
handling the output and calling callback functions. We would cleanup
the processes but not the threads, leading to unprocessed output.

There was also a bug in handle_background_process when trying to
enter the "with" block when the logfile is None.