Fix read_subprocess_pipe bug introduced in 5ddcac2
Calling readlines() on the stream ensures all output is processed,
but it's a blocking op. As a result, we couldn't receive the
notification about server readiness when running the TCP benchmark
from the CLI. This fix continues to read only single lines when
select indicates a stream is ready, but still calls readlines()
when the process is finished.