Basic client quacker implementation
Adapted from the original sidekick binary. We will probably want to
change the interface for proxy configuration from the CLI. It's also
unfortunate we reuse the socket code in the proxy implementation, but
anyway let's not worry too much about it.

Not sure if there would be synchronization issues since the sniffer is
not embedded directly in the client.

For the proxy implementation:
- Without a fancier discovery mechanism, the proxy should assume that
  sniffed UDP packets destined to the proxy IP at the special port 5252
  contain a quack in the UDP payload.
- The proxy should send UDP quack reset messages to the same IP/port as
  the sender of these quack packets. In a similar fashion, the quacker
  inspects the dst IP/port of sniffed packets and considers it a quack
  reset if the port (and IP) match its own.
  Note: we can implement these resets later.