Figure out why popen didn't work for picoquic!
stdout in a subprocess is line-buffered when connected to a terminal but
fully buffered when redirected to a pipe (which is the case when using
subprocess.PIPE). Adding `setvbuf(stdout, NULL, _IOLBF, 0);` to
sample_client.c and sample_server.c in picoquic ensured stdout would be
flushed after each newline.

There was also a bug in common.py where we didn't read all possible
ready lines from a ready stream before continuing.