Prepopulate a static ARP table in one hop network
This should solve the problem where the client sends way too many
Discover packets before the proxy reads them. In the tcpdump logs,
I was observing that the client did not receive a response to an
ARP request for the proxy IP, whereas the proxy had responded
multiple times.

In general the network needs some time to setup before executing
something on the hosts, but I don't know how to get around it other than
sleeping a little.